split large fluctuation.py in two
